+++ src/compas_viewer/configurations/controller_config.py
@@ -1,73 +1,103 @@
 from pathlib import Path
 from typing import Dict
-from typing import List
-from typing import Literal
 from typing import TypedDict
+
+from PySide6.QtCore import Qt
 
 from compas_viewer import DATA
 from compas_viewer.configurations import Config
-
-
-class SelectType(TypedDict):
-    """
-    The type template for the select only.
-    """
-
-    mouse: Literal["left", "right", "middle"]
-    multiselect: str
-    deselect: str
+from compas_viewer.utilities import key_mapper
 
 
 class MouseConfigType(TypedDict):
     """
-    The type template for the mouse only.
+    The type template for the mouse-based movement only.
     """
 
-    zoom: Dict[str, str]
-    pan: Dict[str, str]
-    rotate: Dict[str, str]
-    drag_selection: Dict[str, str]
-    drag_deselection: Dict[str, str]
-    select: SelectType
+    mouse: str
+    modifier: str
 
 
-class KeyConfigType(TypedDict):
+class ActionConfigType(TypedDict):
     """
-    The type template for the key only.
+    The type template for the key-based actions only.
     """
 
-    name: str
-    keys: List[str]
+    key: str
+    modifier: str
 
 
 class ControllerConfigType(TypedDict):
     """
-    The type template for the `controller.json`.
+    The type template for the `controller.json`
+    which contains mouse-based movement and key-based actions.
     """
 
-    mouse: MouseConfigType
-    keys: List[KeyConfigType]
+    pan: MouseConfigType
+    rotate: MouseConfigType
+    drag_selection: MouseConfigType
+    drag_deselection: MouseConfigType
+    multiselect: MouseConfigType
+    deselect: MouseConfigType
+    actions: Dict[str, ActionConfigType]
 
 
-class KeyConfig(Config):
+class ActionConfig:
     """
-    The class representation for the key only.
+    The class representation  of the key-based action configuration.
+    The action config contains two elements, "key" and "modifier".
 
     Parameters
     ----------
-    config : :class:`KeyConfigType`
+    config : :class:`ActionConfigType`
         A TypedDict with defined keys and types.
+
+    Attributes
+    ----------
+    config : :class:`ActionConfigType`
+        A TypedDict with defined keys and types.
+    key : :class:`PySide6.QtCore.Qt.Key`
+        The Qt key.
+    modifier : :class:`PySide6.QtCore.Qt.KeyboardModifier`
+        The Qt modifier.
     """
 
-    def __init__(self, config: KeyConfigType) -> None:
-        super(KeyConfig, self).__init__(config)
-        self.name = config["name"]
-        self.keys = config["keys"]
+    def __init__(self, config: ActionConfigType):
+        self.config = config
+        self.key = key_mapper(config["key"], 0)
+        self.modifier = key_mapper(config["modifier"], 1)
+
+
+class MouseConfig:
+    """
+    The class representation of the mouse-based movement configuration.
+    The mouse contains two elements, "mouse" and "modifier".
+
+    Parameters
+    ----------
+    config : :class:`MouseConfigType`
+        A TypedDict with defined keys and types.
+
+    Attributes
+    ----------
+    config : :class:`MouseConfigType`
+        A TypedDict with defined keys and types.
+    mouse : :class:`PySide6.QtCore.Qt.MouseButton`
+        The Qt mouse.
+    modifier : :class:`PySide6.QtCore.Qt.KeyboardModifier`
+        The Qt modifier.
+    """
+
+    def __init__(self, config: MouseConfigType):
+        self.config = config
+        self.mouse = key_mapper(config["mouse"], 2)
+        self.modifier = key_mapper(config["modifier"], 1)
 
 
 class ControllerConfig(Config):
     """
-    The class representation for the `controller.json` of the class :class:`compas_viewer.controller.Controller`
+    The class representation for the `controller.json` of
+    the class :class:`compas_viewer.controller.controller.Controller`
     The controller.json contains all the settings about controlling the viewer: mouse, keys, ...
 
     Parameters
@@ -78,16 +108,17 @@ class ControllerConfig(Config):
 
     def __init__(self, config: ControllerConfigType) -> None:
         super(ControllerConfig, self).__init__(config)
-        self.pan = config["mouse"]["pan"]
-        self.zoom = config["mouse"]["zoom"]
-        self.rotate = config["mouse"]["rotate"]
-        self.drag_selection = config["mouse"]["drag_selection"]
-        self.drag_deselection = config["mouse"]["drag_deselection"]
-        self.select: str = config["mouse"]["select"]["mouse"]
-        self.multiselect: str = config["mouse"]["select"]["multiselect"]
-        self.deselect: str = config["mouse"]["select"]["deselect"]
-        for key in config["keys"]:
-            setattr(self, key["name"], KeyConfig(key))
+        # Zoom function fixed.
+        self.zoom = None
+        # Select function fixed.
+        self.select = Qt.MouseButton.LeftButton
+        self.pan = MouseConfig(config["pan"])
+        self.rotate = MouseConfig(config["rotate"])
+        self.drag_selection = MouseConfig(config["drag_selection"])
+        self.drag_deselection = MouseConfig(config["drag_deselection"])
+        self.multiselect = MouseConfig(config["multiselect"])
+        self.deselect = MouseConfig(config["deselect"])
+        self.actions = {k: ActionConfig(v) for k, v in config["actions"].items()}
 
     @classmethod
     def from_default(cls) -> "ControllerConfig":