+++ src/compas_viewer/utilities/__init__.py
@@ -1,2 +1,2 @@
 from .gl import gl_info, make_vertex_buffer, make_index_buffer, update_vertex_buffer, update_index_buffer  # noqa: F401
-from .keys import supported_keys  # noqa: F401
+from .qt import key_mapper  # noqa: F401